Rename sub-component
